--- a/requirements/basetest.txt
+++ b/requirements/basetest.txt
@@ -5,27 +5,31 @@
 #
 #    pip-compile-multi
 #
-asttokens==2.4.1
+asttokens==3.0.0
     # via stack-data
-certifi==2024.8.30
+certifi==2024.12.14
     # via requests
-charset-normalizer==3.3.2
+charset-normalizer==3.4.1
     # via requests
 comm==0.2.2
     # via ipywidgets
-contourpy==1.3.0
+contourpy==1.3.1
     # via matplotlib
 cycler==0.12.1
     # via matplotlib
 decorator==5.1.1
     # via ipython
-executing==2.1.0
+exceptiongroup==1.2.2
+    # via
+    #   ipython
+    #   pytest
+executing==2.2.0
     # via stack-data
-fonttools==4.53.1
+fonttools==4.55.6
     # via matplotlib
 graphviz==0.20.3
     # via -r basetest.in
-h5py==3.11.0
+h5py==3.12.1
     # via -r basetest.in
 idna==3.10
     # via requests
@@ -33,29 +37,27 @@ iniconfig==2.0.0
     # via pytest
 ipydatawidgets==4.3.5
     # via pythreejs
-ipympl==0.9.4
+ipympl==0.9.6
     # via -r basetest.in
-ipython==8.27.0
+ipython==8.31.0
     # via
     #   ipympl
     #   ipywidgets
-ipython-genutils==0.2.0
-    # via ipympl
 ipywidgets==8.1.5
     # via
     #   -r basetest.in
     #   ipydatawidgets
     #   ipympl
     #   pythreejs
-jedi==0.19.1
+jedi==0.19.2
     # via ipython
 jupyterlab-widgets==3.0.13
     # via ipywidgets
 kaleido==0.2.1
     # via -r basetest.in
-kiwisolver==1.4.7
+kiwisolver==1.4.8
     # via matplotlib
-matplotlib==3.9.2
+matplotlib==3.10.0
     # via
     #   ipympl
     #   mpltoolbox
@@ -63,7 +65,7 @@ matplotlib-inline==0.1.7
     # via ipython
 mpltoolbox==24.5.1
     # via -r basetest.in
-numpy==2.1.1
+numpy==2.2.2
     # via
     #   contourpy
     #   h5py
@@ -72,19 +74,18 @@ numpy==2.1.1
     #   matplotlib
     #   mpltoolbox
     #   pandas
-    #   pyarrow
     #   pythreejs
     #   scipp
     #   scipy
     #   xarray
-packaging==24.1
+packaging==24.2
     # via
     #   matplotlib
     #   plotly
     #   pooch
     #   pytest
     #   xarray
-pandas==2.2.2
+pandas==2.2.3
     # via
     #   -r basetest.in
     #   xarray
@@ -92,7 +93,7 @@ parso==0.8.4
     # via jedi
 pexpect==4.9.0
     # via ipython
-pillow==10.4.0
+pillow==11.1.0
     # via
     #   ipympl
     #   matplotlib
@@ -104,19 +105,19 @@ pluggy==1.5.0
     # via pytest
 pooch==1.8.2
     # via -r basetest.in
-prompt-toolkit==3.0.47
+prompt-toolkit==3.0.50
     # via ipython
 ptyprocess==0.7.0
     # via pexpect
 pure-eval==0.2.3
     # via stack-data
-pyarrow==17.0.0
+pyarrow==19.0.0
     # via -r basetest.in
-pygments==2.18.0
+pygments==2.19.1
     # via ipython
-pyparsing==3.1.4
+pyparsing==3.2.1
     # via matplotlib
-pytest==8.3.3
+pytest==8.3.4
     # via -r basetest.in
 python-dateutil==2.9.0.post0
     # via
@@ -130,16 +131,16 @@ requests==2.32.3
     # via pooch
 scipp==25.1.0
     # via -r basetest.in
-scipy==1.14.1
+scipy==1.15.1
     # via -r basetest.in
-six==1.16.0
-    # via
-    #   asttokens
-    #   python-dateutil
+six==1.17.0
+    # via python-dateutil
 stack-data==0.6.3
     # via ipython
 tenacity==9.0.0
     # via plotly
+tomli==2.2.1
+    # via pytest
 traitlets==5.14.3
     # via
     #   comm
@@ -151,13 +152,15 @@ traitlets==5.14.3
     #   traittypes
 traittypes==0.2.1
     # via ipydatawidgets
-tzdata==2024.1
+typing-extensions==4.12.2
+    # via ipython
+tzdata==2025.1
     # via pandas
-urllib3==2.2.3
+urllib3==2.3.0
     # via requests
 wcwidth==0.2.13
     # via prompt-toolkit
 widgetsnbextension==4.0.13
     # via ipywidgets
-xarray==2024.9.0
+xarray==2025.1.1
     # via -r basetest.in