--- a/requirements/dev.txt
+++ b/requirements/dev.txt
@@ -14,7 +14,7 @@
 -r wheels.txt
 annotated-types==0.7.0
     # via pydantic
-anyio==4.4.0
+anyio==4.8.0
     # via
     #   httpx
     #   jupyter-server
@@ -28,13 +28,13 @@ async-lru==2.0.4
     # via jupyterlab
 cffi==1.17.1
     # via argon2-cffi-bindings
-click==8.1.7
+click==8.1.8
     # via
     #   pip-compile-multi
     #   pip-tools
-copier==9.3.1
+copier==9.4.1
     # via -r dev.in
-dunamai==1.22.0
+dunamai==1.23.0
     # via copier
 fqdn==1.5.1
     # via jsonschema
@@ -42,15 +42,15 @@ funcy==2.0
     # via copier
 h11==0.14.0
     # via httpcore
-httpcore==1.0.5
+httpcore==1.0.7
     # via httpx
-httpx==0.27.2
+httpx==0.28.1
     # via jupyterlab
 isoduration==20.11.0
     # via jsonschema
 jinja2-ansible-filters==1.3.2
     # via copier
-json5==0.9.25
+json5==0.10.0
     # via jupyterlab-server
 jsonpointer==3.0.0
     # via jsonschema
@@ -59,11 +59,11 @@ jsonschema[format-nongpl]==4.23.0
     #   jupyter-events
     #   jupyterlab-server
     #   nbformat
-jupyter-events==0.10.0
+jupyter-events==0.11.0
     # via jupyter-server
 jupyter-lsp==2.2.5
     # via jupyterlab
-jupyter-server==2.14.2
+jupyter-server==2.15.0
     # via
     #   jupyter-lsp
     #   jupyterlab
@@ -71,7 +71,7 @@ jupyter-server==2.14.2
     #   notebook-shim
 jupyter-server-terminals==0.5.3
     # via jupyter-server
-jupyterlab==4.2.5
+jupyterlab==4.3.4
     # via -r dev.in
 jupyterlab-server==2.27.3
     # via jupyterlab
@@ -81,23 +81,23 @@ overrides==7.7.0
     # via jupyter-server
 pathspec==0.12.1
     # via copier
-pip-compile-multi==2.6.4
+pip-compile-multi==2.7.1
     # via -r dev.in
 pip-tools==7.4.1
     # via pip-compile-multi
-plumbum==1.8.3
+plumbum==1.9.0
     # via copier
-prometheus-client==0.20.0
+prometheus-client==0.21.1
     # via jupyter-server
 pycparser==2.22
     # via cffi
-pydantic==2.9.2
+pydantic==2.10.6
     # via copier
-pydantic-core==2.23.4
+pydantic-core==2.27.2
     # via pydantic
-python-json-logger==2.0.7
+python-json-logger==3.2.1
     # via jupyter-events
-questionary==1.10.0
+questionary==2.1.0
     # via copier
 rfc3339-validator==0.1.4
     # via
@@ -110,24 +110,22 @@ rfc3986-validator==0.1.1
 send2trash==1.8.3
     # via jupyter-server
 sniffio==1.3.1
-    # via
-    #   anyio
-    #   httpx
+    # via anyio
 terminado==0.18.1
     # via
     #   jupyter-server
     #   jupyter-server-terminals
 toposort==1.10
     # via pip-compile-multi
-types-python-dateutil==2.9.0.20240906
+types-python-dateutil==2.9.0.20241206
     # via arrow
 uri-template==1.3.0
     # via jsonschema
-webcolors==24.8.0
+webcolors==24.11.1
     # via jsonschema
 websocket-client==1.8.0
     # via jupyter-server
-wheel==0.44.0
+wheel==0.45.1
     # via pip-tools
 
 # The following packages are considered to be unsafe in a requirements file: