--- a/requirements/docs.txt
+++ b/requirements/docs.txt
@@ -11,7 +11,9 @@ accessible-pygments==0.0.5
     # via pydata-sphinx-theme
 alabaster==1.0.0
     # via sphinx
-attrs==24.2.0
+appnope==0.1.4
+    # via ipykernel
+attrs==25.1.0
     # via
     #   jsonschema
     #   referencing
@@ -23,9 +25,9 @@ beautifulsoup4==4.12.3
     # via
     #   nbconvert
     #   pydata-sphinx-theme
-bleach==6.1.0
+bleach[css]==6.2.0
     # via nbconvert
-debugpy==1.8.5
+debugpy==1.8.12
     # via ipykernel
 defusedxml==0.7.1
     # via nbconvert
@@ -35,17 +37,17 @@ docutils==0.21.2
     #   nbsphinx
     #   pydata-sphinx-theme
     #   sphinx
-fastjsonschema==2.20.0
+fastjsonschema==2.21.1
     # via nbformat
-gitdb==4.0.11
+gitdb==4.0.12
     # via gitpython
-gitpython==3.1.43
+gitpython==3.1.44
     # via -r docs.in
 imagesize==1.4.1
     # via sphinx
 ipykernel==6.29.5
     # via -r docs.in
-jinja2==3.1.4
+jinja2==3.1.5
     # via
     #   myst-parser
     #   nbconvert
@@ -53,7 +55,7 @@ jinja2==3.1.4
     #   sphinx
 jsonschema==4.23.0
     # via nbformat
-jsonschema-specifications==2023.12.1
+jsonschema-specifications==2024.10.1
     # via jsonschema
 jupyter-client==8.6.3
     # via
@@ -72,7 +74,7 @@ markdown-it-py==3.0.0
     # via
     #   mdit-py-plugins
     #   myst-parser
-markupsafe==2.1.5
+markupsafe==3.0.2
     # via
     #   jinja2
     #   nbconvert
@@ -80,28 +82,28 @@ mdit-py-plugins==0.4.2
     # via myst-parser
 mdurl==0.1.2
     # via markdown-it-py
-mistune==3.0.2
+mistune==3.1.0
     # via nbconvert
 myst-parser==4.0.0
     # via -r docs.in
-nbclient==0.10.0
+nbclient==0.10.2
     # via nbconvert
-nbconvert==7.16.4
+nbconvert==7.16.6
     # via nbsphinx
 nbformat==5.10.4
     # via
     #   nbclient
     #   nbconvert
     #   nbsphinx
-nbsphinx==0.9.5
+nbsphinx==0.9.6
     # via -r docs.in
 nest-asyncio==1.6.0
     # via ipykernel
 pandocfilters==1.5.1
     # via nbconvert
-psutil==6.0.0
+psutil==6.1.1
     # via ipykernel
-pydata-sphinx-theme==0.15.4
+pydata-sphinx-theme==0.16.1
     # via -r docs.in
 pyyaml==6.0.2
     # via myst-parser
@@ -109,21 +111,21 @@ pyzmq==26.2.0
     # via
     #   ipykernel
     #   jupyter-client
-referencing==0.35.1
+referencing==0.36.2
     # via
     #   jsonschema
     #   jsonschema-specifications
-rpds-py==0.20.0
+rpds-py==0.22.3
     # via
     #   jsonschema
     #   referencing
-smmap==5.0.1
+smmap==5.0.2
     # via gitdb
 snowballstemmer==2.2.0
     # via sphinx
 soupsieve==2.6
     # via beautifulsoup4
-sphinx==8.0.2
+sphinx==8.1.3
     # via
     #   -r docs.in
     #   myst-parser
@@ -133,13 +135,13 @@ sphinx==8.0.2
     #   sphinx-copybutton
     #   sphinx-design
     #   sphinx-gallery
-sphinx-autodoc-typehints==2.4.1
+sphinx-autodoc-typehints==3.0.1
     # via -r docs.in
 sphinx-copybutton==0.5.2
     # via -r docs.in
 sphinx-design==0.6.1
     # via -r docs.in
-sphinx-gallery==0.17.1
+sphinx-gallery==0.18.0
     # via -r docs.in
 sphinxcontrib-applehelp==2.0.0
     # via sphinx
@@ -153,9 +155,9 @@ sphinxcontrib-qthelp==2.0.0
     # via sphinx
 sphinxcontrib-serializinghtml==2.0.0
     # via sphinx
-tinycss2==1.3.0
-    # via nbconvert
-tornado==6.4.1
+tinycss2==1.4.0
+    # via bleach
+tornado==6.4.2
     # via
     #   ipykernel
     #   jupyter-client