--- a/requirements/mypy.txt
+++ b/requirements/mypy.txt
@@ -6,7 +6,7 @@
 #    pip-compile-multi
 #
 -r test.txt
-mypy==1.11.2
+mypy==1.14.1
     # via -r mypy.in
 mypy-extensions==1.0.0
     # via mypy