--- a/requirements/static.txt
+++ b/requirements/static.txt
@@ -7,19 +7,19 @@
 #
 cfgv==3.4.0
     # via pre-commit
-distlib==0.3.8
+distlib==0.3.9
     # via virtualenv
-filelock==3.16.1
+filelock==3.17.0
     # via virtualenv
-identify==2.6.1
+identify==2.6.6
     # via pre-commit
 nodeenv==1.9.1
     # via pre-commit
 platformdirs==4.3.6
     # via virtualenv
-pre-commit==3.8.0
+pre-commit==4.1.0
     # via -r static.in
 pyyaml==6.0.2
     # via pre-commit
-virtualenv==20.26.4
+virtualenv==20.29.1
     # via pre-commit