--- a/requirements/wheels.txt
+++ b/requirements/wheels.txt
@@ -5,11 +5,11 @@
 #
 #    pip-compile-multi
 #
-build==1.2.2
+build==1.2.2.post1
     # via -r wheels.in
-packaging==24.1
+packaging==24.2
     # via build
-pyproject-hooks==1.1.0
+pyproject-hooks==1.2.0
     # via build
-tomli==2.0.1
+tomli==2.2.1
     # via build