--- a/src/plopp/core/__init__.pyi
+++ b/src/plopp/core/__init__.pyi
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
+
+from .graph import show_graph
+from .helpers import node, widget_node
+from .node_class import Node
+from .view import View
+
+__all__ = ['Node', 'View', 'node', 'show_graph', 'widget_node']