--- a/src/plopp/graphics/__init__.py
+++ b/src/plopp/graphics/__init__.py
@@ -1,31 +1,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
+# Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
 
-from .basefig import BaseFig
-from .bbox import BoundingBox
-from .camera import Camera
-from .colormapper import ColorMapper
-from .figures import (
-    linefigure,
-    imagefigure,
-    mesh3dfigure,
-    scatterfigure,
-    scatter3dfigure,
-)
-from .graphicalview import GraphicalView
-from .tiled import tiled
+import lazy_loader as lazy
 
-
-__all__ = [
-    'BaseFig',
-    'BoundingBox',
-    'Camera',
-    'ColorMapper',
-    'GraphicalView',
-    'imagefigure',
-    'linefigure',
-    'mesh3dfigure',
-    'scatter3dfigure',
-    'scatterfigure',
-    'tiled',
-]
+__getattr__, __dir__, __all__ = lazy.attach_stub(__name__, __file__)