--- a/src/plopp/graphics/__init__.pyi
+++ b/src/plopp/graphics/__init__.pyi
@@ -1,0 +1,30 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
+
+from .basefig import BaseFig
+from .bbox import BoundingBox
+from .camera import Camera
+from .colormapper import ColorMapper
+from .figures import (
+    imagefigure,
+    linefigure,
+    mesh3dfigure,
+    scatter3dfigure,
+    scatterfigure,
+)
+from .graphicalview import GraphicalView
+from .tiled import tiled
+
+__all__ = [
+    'BaseFig',
+    'BoundingBox',
+    'Camera',
+    'ColorMapper',
+    'GraphicalView',
+    'imagefigure',
+    'linefigure',
+    'mesh3dfigure',
+    'scatter3dfigure',
+    'scatterfigure',
+    'tiled',
+]