--- a/src/plopp/plotting/__init__.py
+++ b/src/plopp/plotting/__init__.py
@@ -1,22 +1,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
+# Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
 
-from .inspector import inspector
-from .mesh3d import mesh3d
-from .plot import plot
-from .scatter import scatter
-from .scatter3d import scatter3d
-from .slicer import slicer
-from .superplot import superplot
-from .xyplot import xyplot
+import lazy_loader as lazy
 
-__all__ = [
-    'inspector',
-    'mesh3d',
-    'plot',
-    'scatter',
-    'scatter3d',
-    'slicer',
-    'superplot',
-    'xyplot',
-]
+__getattr__, __dir__, __all__ = lazy.attach_stub(__name__, __file__)