--- a/src/plopp/plotting/__init__.pyi
+++ b/src/plopp/plotting/__init__.pyi
@@ -1,0 +1,22 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
+
+from .inspector import inspector
+from .mesh3d import mesh3d
+from .plot import plot
+from .scatter import scatter
+from .scatter3d import scatter3d
+from .slicer import slicer
+from .superplot import superplot
+from .xyplot import xyplot
+
+__all__ = [
+    'inspector',
+    'mesh3d',
+    'plot',
+    'scatter',
+    'scatter3d',
+    'slicer',
+    'superplot',
+    'xyplot',
+]