--- a/src/plopp/widgets/__init__.py
+++ b/src/plopp/widgets/__init__.py
@@ -1,32 +1,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
+# Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
 
-from .box import Box, HBar, VBar
-from .checkboxes import Checkboxes
-from .clip3d import Clip3dTool, ClippingPlanes
-from .drawing import DrawingTool, PointsTool
-from .linesave import LineSaveTool
-from .slice import RangeSliceWidget, SliceWidget, slice_dims
-from .toolbar import Toolbar, make_toolbar_canvas2d, make_toolbar_canvas3d
-from .tools import ButtonTool, ColorTool, ToggleTool
+import lazy_loader as lazy
 
-__all__ = [
-    "Box",
-    "ButtonTool",
-    "Checkboxes",
-    "Clip3dTool",
-    "ClippingPlanes",
-    "ColorTool",
-    "DrawingTool",
-    "HBar",
-    "LineSaveTool",
-    "PointsTool",
-    "RangeSliceWidget",
-    "SliceWidget",
-    "ToggleTool",
-    "Toolbar",
-    "VBar",
-    "make_toolbar_canvas2d",
-    "make_toolbar_canvas3d",
-    "slice_dims",
-]
+__getattr__, __dir__, __all__ = lazy.attach_stub(__name__, __file__)