--- a/src/plopp/widgets/__init__.pyi
+++ b/src/plopp/widgets/__init__.pyi
@@ -1,0 +1,32 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
+
+from .box import Box, HBar, VBar
+from .checkboxes import Checkboxes
+from .clip3d import Clip3dTool, ClippingPlanes
+from .drawing import DrawingTool, PointsTool
+from .linesave import LineSaveTool
+from .slice import RangeSliceWidget, SliceWidget, slice_dims
+from .toolbar import Toolbar, make_toolbar_canvas2d, make_toolbar_canvas3d
+from .tools import ButtonTool, ColorTool, ToggleTool
+
+__all__ = [
+    "Box",
+    "ButtonTool",
+    "Checkboxes",
+    "Clip3dTool",
+    "ClippingPlanes",
+    "ColorTool",
+    "DrawingTool",
+    "HBar",
+    "LineSaveTool",
+    "PointsTool",
+    "RangeSliceWidget",
+    "SliceWidget",
+    "ToggleTool",
+    "Toolbar",
+    "VBar",
+    "make_toolbar_canvas2d",
+    "make_toolbar_canvas3d",
+    "slice_dims",
+]