… requirements/base.in
… requirements/base.txt
… requirements/nightly.in
… requirements/nightly.txt
… src/plopp/core/node_class.py