In the spirit of https://github.com/scipp/scippneutron/pull/596, this adds lazy_loader to plopp.
Import time goes from 

```
In [1]: %time import plopp
CPU times: user 234 ms, sys: 33.7 ms, total: 268 ms
Wall time: 343 ms
```
to
```
In [1]: %time import plopp
CPU times: user 3.52 ms, sys: 3.61 ms, total: 7.12 ms
Wall time: 7.15 ms
```

(not a very scientific comparison but still seems like a decent speedup).

The `attach_stub` function works by parsing the `.pyi` files to infer submodules and functions to expose. It also makes sure that static typecheckers and IDE don't complain too much.